Move prod types to EssentialFeed Prod modules
